requirements.txt
beautifulsoup4==4.8.2
certifi==2020.6.20
chardet==3.0.4
Django==2.2.11
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
django-crispy-forms==1.9.0
django-stdimage==5.0.2
idna==2.10
Pillow==7.0.0
progressbar2==3.50.1
psycopg2==2.8.4
python-utils==2.4.0
pytz==2019.3
requests==2.24.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
stripe==2.50.0
urllib3==1.25.10
